网盘 V2 V3 兼容 问题 图片保存到相册功能

+++ app/src/main/java/net/zoneland/x/bpm/mobile/v1/zoneXBPM/app/clouddrive/v2/f/FileFolderListPresenter.kt
@@ -1,6 +1,7 @@
 package net.zoneland.x.bpm.mobile.v1.zoneXBPM.app.clouddrive.v2.f
 
 import net.zoneland.x.bpm.mobile.v1.zoneXBPM.O2
+import net.zoneland.x.bpm.mobile.v1.zoneXBPM.O2SDKManager
 import net.zoneland.x.bpm.mobile.v1.zoneXBPM.app.base.BasePresenterImpl
 import net.zoneland.x.bpm.mobile.v1.zoneXBPM.model.bo.api.ApiResponse
 import net.zoneland.x.bpm.mobile.v1.zoneXBPM.model.bo.api.IdData
@@ -23,26 +24,113 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
 
     override fun move(files: List<CloudDiskItem.FileItem>, folders: List<CloudDiskItem.FolderItem>, destFolderId: String) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
         val all : ArrayList<Observable<ApiResponse<IdData>>> = ArrayList()
-        if (files.isNotEmpty()) {
-            all.addAll(
+        if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            val service = getCloudFileV3ControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            if (files.isNotEmpty()) {
+                all.addAll(
                     files.map {
-                        val json = FileJson(it.id, it.createTime, it.updateTime, it.name, it.person,
-                                "", it.fileName, it.extension, it.contentType, it.storageName, it.fileId,
-                                it.storage, it.type, it.length, destFolderId, it.lastUpdateTime, it .lastUpdatePerson)
+                        val json = FileJson(
+                            it.id,
+                            it.createTime,
+                            it.updateTime,
+                            it.name,
+                            it.person,
+                            "",
+                            it.fileName,
+                            it.extension,
+                            it.contentType,
+                            it.storageName,
+                            it.fileId,
+                            it.storage,
+                            it.type,
+                            it.length,
+                            destFolderId,
+                            it.lastUpdateTime,
+                            it.lastUpdatePerson
+                        )
                         service.updateFile(json, it.id)
                     }
-            )
-        }
-        if (folders.isNotEmpty()) {
-            all.addAll(
+                )
+            }
+            if (folders.isNotEmpty()) {
+                all.addAll(
                     folders.map {
-                        val json = FolderJson(it.id, it.createTime, it.updateTime, it.name, it.person,
-                                "", destFolderId, it.attachmentCount, it.size, it.folderCount, it.status, it.fileId)
+                        val json = FolderJson(
+                            it.id,
+                            it.createTime,
+                            it.updateTime,
+                            it.name,
+                            it.person,
+                            "",
+                            destFolderId,
+                            it.attachmentCount,
+                            it.size,
+                            it.folderCount,
+                            it.status,
+                            it.fileId
+                        )
                         service.updateFolder(it.id, json)
                     }
-            )
+                )
+            }
+        } else {
+            val service = getCloudFileControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            if (files.isNotEmpty()) {
+                all.addAll(
+                    files.map {
+                        val json = FileJson(
+                            it.id,
+                            it.createTime,
+                            it.updateTime,
+                            it.name,
+                            it.person,
+                            "",
+                            it.fileName,
+                            it.extension,
+                            it.contentType,
+                            it.storageName,
+                            it.fileId,
+                            it.storage,
+                            it.type,
+                            it.length,
+                            destFolderId,
+                            it.lastUpdateTime,
+                            it.lastUpdatePerson
+                        )
+                        service.updateFile(json, it.id)
+                    }
+                )
+            }
+            if (folders.isNotEmpty()) {
+                all.addAll(
+                    folders.map {
+                        val json = FolderJson(
+                            it.id,
+                            it.createTime,
+                            it.updateTime,
+                            it.name,
+                            it.person,
+                            "",
+                            destFolderId,
+                            it.attachmentCount,
+                            it.size,
+                            it.folderCount,
+                            it.status,
+                            it.fileId
+                        )
+                        service.updateFolder(it.id, json)
+                    }
+                )
+            }
         }
         Observable.merge(all).subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
@@ -58,14 +146,35 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
     }
 
+
     override fun share(ids: List<String>, users: List<String>, orgs: List<String>) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
-        val all = ids.map {
-            val form = CloudDiskShareForm()
-            form.fileId = it
-            form.shareOrgList = orgs
-            form.shareUserList = users
-            service.share(form)
+        val all : ArrayList<Observable<ApiResponse<IdData>>> = ArrayList()
+        if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            val service = getCloudFileV3ControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            all.addAll(ids.map {
+                val form = CloudDiskShareForm()
+                form.fileId = it
+                form.shareOrgList = orgs
+                form.shareUserList = users
+                service.share(form)
+            })
+        } else {
+            val service = getCloudFileControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            all.addAll(ids.map {
+                val form = CloudDiskShareForm()
+                form.fileId = it
+                form.shareOrgList = orgs
+                form.shareUserList = users
+                service.share(form)
+            })
         }
         Observable.merge(all).subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
@@ -83,16 +192,37 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
 
     override fun deleteBatch(fileIds: List<String>, folderIds: List<String>) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
         val all : ArrayList<Observable<ApiResponse<IdData>>> = ArrayList()
-        if (fileIds.isNotEmpty()) {
-            val deletes = fileIds.map { service.deleteFile(it) }
-            all.addAll(deletes)
+        if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            val service = getCloudFileV3ControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            if (fileIds.isNotEmpty()) {
+                val deletes = fileIds.map { service.deleteFile(it) }
+                all.addAll(deletes)
+            }
+            if (folderIds.isNotEmpty()) {
+                val deleteFolders = folderIds.map { service.deleteFolder(it) }
+                all.addAll(deleteFolders)
+            }
+        } else {
+            val service = getCloudFileControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            if (fileIds.isNotEmpty()) {
+                val deletes = fileIds.map { service.deleteFile(it) }
+                all.addAll(deletes)
+            }
+            if (folderIds.isNotEmpty()) {
+                val deleteFolders = folderIds.map { service.deleteFolder(it) }
+                all.addAll(deleteFolders)
+            }
         }
-        if (folderIds.isNotEmpty()) {
-            val deleteFolders = folderIds.map { service.deleteFolder(it) }
-            all.addAll(deleteFolders)
-        }
+
         Observable.merge(all).subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .o2Subscribe {
@@ -108,8 +238,16 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
 
     override fun updateFolder(folder: FolderJson) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
-        service.updateFolder(folder.id, folder).subscribeOn(Schedulers.io())
+        val service = if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            getCloudFileV3ControlService(mView?.getContext())?.updateFolder(folder.id, folder)
+        } else {
+            getCloudFileControlService(mView?.getContext())?.updateFolder(folder.id, folder)
+        }
+        if (service == null) {
+            mView?.error("服务为空！")
+            return
+        }
+        service.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .o2Subscribe {
                     onNext {
@@ -124,8 +262,16 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
 
     override fun updateFile(file: FileJson) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
-        service.updateFile(file, file.id).subscribeOn(Schedulers.io())
+        val service = if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            getCloudFileV3ControlService(mView?.getContext())?.updateFile(file, file.id)
+        } else {
+            getCloudFileControlService(mView?.getContext())?.updateFile(file, file.id)
+        }
+        if (service == null) {
+            mView?.error("服务为空！")
+            return
+        }
+        service.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .o2Subscribe {
                     onNext {
@@ -140,15 +286,23 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
 
     override fun uploadFile(parentId: String, file: File) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
         var folderId = parentId
         if (parentId.isEmpty()) {
             folderId = O2.FIRST_PAGE_TAG
         }
         val requestBody = RequestBody.create(MediaType.parse("application/octet-stream"), file)
         val body = MultipartBody.Part.createFormData("file", file.name, requestBody)
-        service.uploadFile2Folder(body, folderId)
-                .subscribeOn(Schedulers.io())
+        val service = if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            getCloudFileV3ControlService(mView?.getContext())?.uploadFile2Folder(body, folderId)
+        } else {
+            getCloudFileControlService(mView?.getContext())?.uploadFile2Folder(body, folderId)
+        }
+        if (service == null) {
+            mView?.error("服务为空！")
+            return
+        }
+
+        service.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .o2Subscribe {
                     onNext {
@@ -200,12 +354,30 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
         }
         val requestBody = RequestBody.create(MediaType.parse("application/octet-stream"), file)
         val body = MultipartBody.Part.createFormData("file", file.name, requestBody)
-        return getCloudFileControlService(mView?.getContext())?.uploadFile2Folder(body, folderId)
+        return if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            getCloudFileV3ControlService(mView?.getContext())?.uploadFile2Folder(
+                body,
+                folderId
+            )
+        } else {
+            getCloudFileControlService(mView?.getContext())?.uploadFile2Folder(
+                body,
+                folderId
+            )
+        }
     }
 
     override fun createFolder(params: HashMap<String, String>) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
-        service.createFolder(params).subscribeOn(Schedulers.io())
+        val service = if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            getCloudFileV3ControlService(mView?.getContext())?.createFolder(params)
+        } else {
+            getCloudFileControlService(mView?.getContext())?.createFolder(params)
+        }
+        if (service == null) {
+            mView?.error("服务为空！")
+            return
+        }
+        service.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .o2Subscribe {
                     onNext {
@@ -220,67 +392,60 @@ class FileFolderListPresenter : BasePresenterImpl<FileFolderListContract.View>()
 
 
     override fun getItemList(parentId: String) {
-        val service = getCloudFileControlService(mView?.getContext()) ?: return
-        if (parentId.isEmpty()) {
-            Observable
-                    .zip(service.listFolderTop(), service.listFileTop()) { r1, r2 ->
-                        val list = ArrayList<CloudDiskItem>()
-                        val folderList = r1.data
-                        val fileList = r2.data
-                        if (folderList != null && folderList.isNotEmpty()) {
-                            folderList.forEach {
-                                list.add(it.copyToVO2())
-                            }
-                        }
-                        if (fileList!=null && fileList.isNotEmpty()) {
-                            fileList.forEach {
-                                list.add(it.copyToVO2())
-                            }
-                        }
-                        list
-                    }
-                    .subscribeOn(Schedulers.io())
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .o2Subscribe {
-                        onNext {
-                            mView?.itemList(it)
-                        }
-                        onError { e, _ ->
-                            XLog.error("", e)
-                            mView?.error(e?.message ?: "错误！")
-                        }
-                    }
-        }else {
-            Observable
-                    .zip(service.listFolderByFolderId(parentId), service.listFileByFolderId(parentId)) { r1, r2 ->
-                        val list = ArrayList<CloudDiskItem>()
-                        val folderList = r1.data
-                        val fileList = r2.data
-                        if (folderList != null && folderList.isNotEmpty()) {
-                            folderList.forEach {
-                                list.add(it.copyToVO2())
-                            }
-                        }
-                        if (fileList!=null && fileList.isNotEmpty()) {
-                            fileList.forEach {
-                                list.add(it.copyToVO2())
-                            }
-                        }
-                        list
-                    }
-                    .subscribeOn(Schedulers.io())
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .o2Subscribe {
-                        onNext {
-                            mView?.itemList(it)
-                        }
-                        onError { e, _ ->
-                            XLog.error("", e)
-                            mView?.error(e?.message ?: "错误！")
-                        }
-                    }
+        if (O2SDKManager.instance().appCloudDiskIsV3()) {
+            val serviceV3 = getCloudFileV3ControlService(mView?.getContext())
+            if (serviceV3 == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            if (parentId.isEmpty()) {
+                getItemListObservable(serviceV3.listFolderTop(), serviceV3.listFileTop())
+            } else {
+                getItemListObservable(serviceV3.listFolderByFolderId(parentId), serviceV3.listFileByFolderId(parentId))
+            }
+        } else {
+            val service = getCloudFileControlService(mView?.getContext())
+            if (service == null) {
+                mView?.error("服务为空！")
+                return
+            }
+            if (parentId.isEmpty()) {
+                getItemListObservable(service.listFolderTop(), service.listFileTop())
+            } else {
+                getItemListObservable(service.listFolderByFolderId(parentId), service.listFileByFolderId(parentId))
+            }
         }
+    }
 
+    private fun getItemListObservable(folderObservable: Observable<ApiResponse<List<FolderJson>>>, fileObservable: Observable<ApiResponse<List<FileJson>>>) {
+        Observable
+            .zip(folderObservable, fileObservable) { r1, r2 ->
+                val list = ArrayList<CloudDiskItem>()
+                val folderList = r1.data
+                val fileList = r2.data
+                if (folderList != null && folderList.isNotEmpty()) {
+                    folderList.forEach {
+                        list.add(it.copyToVO2())
+                    }
+                }
+                if (fileList!=null && fileList.isNotEmpty()) {
+                    fileList.forEach {
+                        list.add(it.copyToVO2())
+                    }
+                }
+                list
+            }
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
+            .o2Subscribe {
+                onNext {
+                    mView?.itemList(it)
+                }
+                onError { e, _ ->
+                    XLog.error("", e)
+                    mView?.error(e?.message ?: "错误！")
+                }
+            }
     }
 
 }
